RKUSBDriver: allow to specify the offset at which the image should be flashed
RKUSBDriver flashes an image onto a storage medium. Typically, on eMMC
the starting offset is 0x40 (i.e. 64 blocks of 512KiB, iow 32KiB) to
leave enough room for a partition table before but it doesn't *have to*.
Barebox's update mechanism (bbu/CONFIG_BAREBOX_UPDATE) specifies
multiple possible offsets for Rockchip: 0x40 to 0x44[1].

Additionally, it is possible (on RK3399 at the very least) to have the
first stage bootloader at offset 0 on SPI-NOR flashes. This is currently
the choice made for Theobroma Systems's Puma RK3399 SoM in U-Boot.

Let's give the user the ability to provide an offset from the
configuration file.

[1] https://github.com/barebox/barebox/blob/master/arch/arm/mach-rockchip/bbu.c#L19-L24
Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>